again a fix for irq handling, fix wrong order of irq execution
If we have a lower level interrupt pending and the I flag is set
we fetch the address of the corrosponding irq vector. During the
pipeline fetch of the irq vector a second higher level irq can be
raised. But we will execute the already fetched lower level irq first!